Allow compound fields to not insert their prefix on the leaves
This way we can create a compound primary key as a nested entity,
but leave the prefix off so the columns in the table are actually
named after the individual leaf-level components.  This is a much
more traditional SQL database design as columns can more easily
share the same name in different tables if they have a foreign
key relationship.

Signed-off-by: Shawn O. Pearce <sop@google.com>