Refactor SQL type generation to allow SqlDialect to override type names
In many cases we don't need to replace the entire SqlTypeInfo object,
or even want to subclass it, when all that needs to change is what
SQL text is used for the name of the column type.  For example, its
a "BLOB" in H2 but "BYTEA" in PostgreSQL for the byte[] column type.

Signed-off-by: Shawn O. Pearce <sop@google.com>