Mostly revert the timestamp muddling done by the prior commit
My prior attempt to fix the timestamp issues I was seeing was
horribly misguided and wrong.

The Timestamp objects have the correct time from the epoch; its
the toString() method that is applying the JVM's time zone to them.
I need to fix the formatter in gwtjsonrpc to format in UTC rather
than assuming toString() will put out a UTC string.

Signed-off-by: Shawn O. Pearce <sop@google.com>