Allow dialects to refine themselves based on connection specific data
Once we have the connection open and have roughly selected the
dialect based on the JDBC URL its useful to allow the dialect
to refine itself to a server-specific version.  This might then
later be used to refine how SQL is generated for that database,
like when syntax is changed or new useful syntax has been added
in later revisions of the server software.

Signed-off-by: Shawn O. Pearce <sop@google.com>