Allow dialects to add their own storage clauses to CREATE TABLE
Some databases permit custom clauses to be inserted after the
closing parenthesis of the CREATE TABLE statement, specifically
to alter how the server stores the data in that table.  As the
syntax is almost always server specific the dialect must have
full control over what is placed here, if anything.

Signed-off-by: Shawn O. Pearce <sop@google.com>