Default our PostgreSQL tables to be WITHOUT OIDS
Its strongly recommended that all user tables in a PostgreSQL schema
be built as "WITHOUT OIDS", or more recently "WITH (OIDS = FALSE)".
This reduces the risk of oid wraparound during normal operation, as
user operations don't cause allocations from the system oid sequence.

Since the syntax changed in PostgreSQL 8.2, we use the new refine
method to select a Pre8.2 dialect for the older "WITHOUT OIDS"
syntax, and the default dialect for 8.2 and newer to use the new
"WITH (OIDS = FALSE)" syntax.

Signed-off-by: Shawn O. Pearce <sop@google.com>