Remove nonsense default value on NOT NULL TIMESTAMP
MySQL doesn't like our phony default value when the local timezone
is GMT+10, apparently that's still a date that is out of range for
the MySQL server.

Instead of registering a phony default which is never used, we
should just leave the default out and require that anyone who does
an INSERT will populate the column.  Certainly any insert done from
the ORM's own code will always populate the column.

Signed-off-by: Shawn O. Pearce <sop@google.com>