Refactor sequence increment to be in dialects
By changing the nextLong method to generate the SQL on the fly within
the increment rather than embedding it into the generated schema
implementation we can later refactor the schema code generator to be
more generic for the NoSQL type of backend where we don't necessarily
have a query statement.  This also cleans up the corner case of
the MySQL dialect where we don't have a single query statement.

Change-Id: Id29923f742c847e16e48655b928ab40f7c827c0f
Signed-off-by: Shawn O. Pearce <sop@google.com>