Refactor SchemaGen to be reused for NoSQL backends
The entire schema implementation is just about constructing the
various methods that return the Access implementations for each
relation.  This isn't unique to JDBC and can easily be reused
for a NoSQL backend as-is, if we properly abstract out the base
class it uses and the method to obtain each Access implementation
as its required by the schema code generator.

Change-Id: If9b1d63b1b83fafb1eab14b64861cd26dde73d2e
Signed-off-by: Shawn O. Pearce <sop@google.com>