Create an encoder for NoSQL secondary index keys
In a NoSQL system we usually don't have a secondary index available
for columns and instead must fake that ourselves by creating a new
row that is a composite of the various columns.  To ensure sorting
occurs properly in the NoSQL key range space, we need to encode the
column values to build the composite key.

Change-Id: Ieefe9757f0005bb55e30c3dc0786769bf2d402fb
Signed-off-by: Shawn O. Pearce <sop@google.com>