Fix protobuf code generation for Timestamp
We forgot to push the field value onto the stack, so we couldn't
invoke getTime() on the timestamp object.  This resulted in a class
verification error.

We also failed to denote the constructor as taking a primitive long,
causing a class verification error when a long was found where an
uninitialized object was expected.

Change-Id: I4c60fac33f2c1c8dea02d97ac6ad4522ffacf3b6
Signed-off-by: Shawn O. Pearce <sop@google.com>