Define a skeleton NoSQL base for extension
This is a simple base that a specific NoSQL implementation can
build on top of as part of our public API.

Trivial implementations are supplied based upon a simple TreeMap in
the JVM, and persisting that map to disk.  These can be useful for
development and testing of applications built on top of the library.

Change-Id: Ida7fdfa4aca5db5cfbc1ca7b7bbec5af70b551fc
Signed-off-by: Shawn O. Pearce <sop@google.com>