Generate a Protobuf definition for the database
I added support for generating a file containing all the Protobuf
type definitions for a given database schema. The JavaSchemaModel now
has a method to output a Protobuf definition to a given PrintWriter.
This allows tools using a NoSQL database to have a Protobuf file
describing the Protobuf data in the database.

Change-Id: I53f4902d2b335be2b20d966fd0eb58c050e8a53c