Fix UPSERT for JDBC batch returning SUCCESS_NO_INFO
On some databases (such as Oracle or MaxDB), the update count array
returned by the Statement.executeBatch may contain the general success
indicator SUCCES_NO_INFO. On such databases a batch UPDATE cannot be
used to check for existence of a row on the database: the situation that
a row can't be updated because it is not existing will also yield
SUCCESS_NO_INFO. For the databases in question, individual updates are
performed with this patch.

Some databases (such as MaxDB) allow to efficiently determine the
overall update count of a batch execution. This patch allows to leverage
this capability in a database-dialect.

Change-Id: Id7cb496fc517e0c31eef2f3c34f566aaac20019d
Signed-off-by: Adrian Goerler <adrian.goerler@sap.com>