Support databases that return SUCCESS_NO_INFO upon batch execution
The PreparedStatement.executeBatch method may return a general success
indicator SUCCESS_NO_INFO. For example, this is the case for Oracle or
SAP MaxDB. If SUCCESS_NO_INFO is returned, the number of database rows
updated or deleted by a particular row of the prepared statement batch
cannot be determined. In particular it is not known whether 0 or 1 rows
have been updated as updating 0 rows is a successful statement execution
as well, from JDBC's point of view.

To perform optimistic concurrency control, it is necessary to determine
whether all items have been successfully updated or deleted. It is not
needed to known which individual item had been modified concurrently. If
batch execution does not allow to determine the overall batch
update count, individual statements need to be executed for
the individual items. This patch delegates the batch execution to the
SQLDialect to allow for a database-vendor-specific implementation.

To implement the upsert functionality, firstly updates are attempted.
For rows that could not be updated successfully, an insert will be
performed. It is necessary to determine the exact update count for
individual items. If this is not possible for a particular SQL dialect,
updates need to be attempted individually.

The SQLDialect is equipped with methods to indicate whether batch
execution allows to determine individual or overall batch update counts.

Change-Id: I779f5a70acbb1886ea7be003a322970f7d7efb0e
Signed-off-by: Adrian Goerler <adrian.goerler@sap.com>