Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This patch introduces an issue: It is not prepared for the fact that GWTORM uses an interface "ResultSet", which extends java.lang.Iterable but does not allow to invoke the iterator() method multiple times, hence violating the Liskov substitution principle. 

I will upload a new change, that takes care of this issue.

Patch-set: 1