Make ResultSet#toList() return a mutable list
Due to some recent changes now the list returned by
ResultSet#toList() may be immuatable (e.g. if an empty list is
returned). This broke some code in Gerrit which assumes that the
returned list is mutable.

To fix this problem this change ensures that ResultSet#toList()
always returns a mutable list by returning a copy of the input. It is
now defined that ResultSet#toList() must return a mutable list.

Change-Id: I720a3f9fc74cc1033cdb403385b2f2816f0f9b64
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>