Allow for "ResultSet"s as arguments of insert, update, delete, upsert
GWTORM defines a "ResulSet" interface, which extends java.lang.Iterator
but does not allow invoking the iterator() method multiple times. This
patch adds a test to assert that implementations of this interface can
be used as arguments for the JdbcAccess.insert, update, delete and
upsert method. It fixes an issue with JdbcAccess.upsert revealed by the
test.

Change-Id: I6ce625ed9476e1649600dc7eb5db4f77e9629607
Signed-off-by: Adrian Goerler <adrian.goerler@sap.com>