Update patch set 1
Patch Set 1:

I am not sure I am getting your point. I am proposing to add an additional launch config for Indigo and keep the existing for Helios as I was under the impression that we are supporting development in Helios as well as in Indigo. If this is not the case, I am equally happy just to change the original launch config to the org.eclipse.m2e namespace.

Which way should we go?

Patch-set: 1