Update patch set 1
Patch Set 1: Do not submit

I am guessing that this change should not be merged.  After talking to rtyler on IRC, he complained about this bug:

  https://gist.github.com/1473386

I happened to remember that I had seen this diff:

  https://gist.github.com/1473583

here:

  https://gerrit-review.googlesource.com/#/c/16562/6/gerrit-server/src/main/java/com/google/gerrit/server/patch/PublishComments.java

I am guessing it was to fix the same problem (rtyler confirms that it does fix it for him).  If so, I suspect that this is as a result of a change in gwtorm?

Either way, this tells me that we should not be releasing the current gwtorm snapshot (or later) as the gwtorm to be included in 2.2.2.  Instead, we should pick a gwtorm version which is based on what Gutasf would have used during his stability testing before rsac went down.  Gustaf, can you let me know what that was?

Perhaps this current version change should indeed become gwtorm2.0, but we should first release a gwtorm 1.6 and build gerrit 2.2.2 off of the new gwtorm 1.6?

Patch-set: 1
Reviewer: Gerrit User 1003883 <1003883@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-2
Label: Verified=0