Update patch set 1
Patch Set 1:

I think that perhaps it is still reasonable to assume that gwtorm returns an immutable List and we should fix Gerrit instead?

Either way, I still think that 2.2.2 should be released with a gwtorm that is close to what SEMC used for their testing, if not we are not benefiting from their testing as this incident highlights.  Do you agree?

Patch-set: 1