Only make primary key protobuf fields required
The primary key is extremely hard to change, so make it required
as the fields must all be not null and populated in order to have
a valid entity.

Other fields may have been added after instances of the entity were
already created. Permit those to be optional, so existing data does
not have to be modified before being used with the new definition.

Change-Id: I1865dffda7a5a4004fd77aa0a37aa1e3ae20976f