Generate AnyFooPrimaryKey protobuf message
This message contains all of the primary key fields of any relation
declared in the schema. It can be useful when trying to name an
entity in a generic way, as any primary key can be wrapped into this
message to denote both the relation and the specific item within it.

Change-Id: Ie8a8a33fc362edf4ff3dc7cced41e5b2851054c4