Bug: OrmDuplicateException isn't raised
When we try to insert a duplicate key, convertError should transform the SqlException to a
OrmDuplicateKeyException. This is the case for PostGreSqlDialect but not for MysqlDialect.
So let's enforce every dialect to implement this feature by making SqlDialect.convertError abstract.

Additionally add a test superclass for asserting error type and message on all derivates of SqlDialect.

Change-Id: Ic5d507e8a4e896a0e038f7109f38013aa2e2c8a1