Update patch set 1
Patch Set 1:

Glad to help.

As for "how did this ever work", the answer is it didn't. This is close to a revert of the parent Ib298edcb0d, but the version number was never bumped, so this never got picked up by a released Gerrit. It only showed up if you built gwtorm from scratch and poisoned your local ~/.m2/repository. Like I did when setting up Eclipse on a brand new machine for GWT debugging.

And now I'm kicking myself for spending hours understanding how gwtorm worked without first bisecting back to the 1.5 release :)

Patch-set: 1