Update patch set 2
Patch Set 2:

Currently, if these databases are not installed or configured, the tests are going to fail. Even worse, a developer may use -DskipTests java system properties disabling even H2 tests, that not needing dedicated database to be installed and configured.

With this change in place instead failing, the tests are skipped. If the prerequisites are met, then the tests are executed.

Patch-set: 2
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0