Update patch set 6
Patch Set 6: Verified-1

The tests added in 78698dbf1a629456fac509a65b37ab565109c412 show that it is broken. The values for short, long and integer are inserted as 0 and overwrite the default value defined in @DefaultValue annotation. I think that generated code in AccessGen.bindOneInsert() has to respect the DefaultValues: either omit the columns with default values from the insert statement entirely or insert it with default values defined in the annotation.

String column is broken too: even if @DefaultValue annotation presents, NULL value is inserted and it fails because the column was defined as NOT NULL.

Patch-set: 6
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=-1