Update patch set 6
Patch Set 6:

So, i think I am missing something. In an ORM default values make only sense when in the class we can let the attributes unset (NULL). This implies that the underlying code actually works with NULL values. But it apparently not, for these types:

* Character
* Boolean
* Short
* Integer
* Long

it fails to execute the select with error in the generated code: default constructor for the classes above not found.

Actually only default value for String type makes currently sense in gwtorm.

Patch-set: 6