Update patch set 3
Patch Set 3:

Have you built it with mysql and postgres tests enabled?

When I run `mvn clean package` it is failing with:

  Failed tests:     
  testListTables(com.google.gwtorm.schema.sql.DialectMySQLTest)
  testListTables(com.google.gwtorm.schema.sql.DialectPostgreSQLTest)

The tagged 1.7 is also failing with the same errors.  It works only if I apply the patches in those test files' setUp methods that were added in [1] and [2].

[1] https://gerrit-review.googlesource.com/#/c/48830/5/src/test/java/com/google/gwtorm/schema/sql/DialectMySQLTest.java

[2] https://gerrit-review.googlesource.com/#/c/48830/5/src/test/java/com/google/gwtorm/schema/sql/DialectPostgreSQLTest.java

Patch-set: 3
Label: Verified=0