Update patch set 4
Patch Set 4:

> Is it expected that the build fails when oracle is not available? 
 > reading the commit message of [1] I understood that it would just
 > skip those tests, but I am getting this:
 > 
 > Tests in error:
 > testUpgradeSchema(com.google.gwtorm.schema.sql.DialectOracleSQLTest):
 > got: <java.lang.ClassNotFoundException: oracle.jdbc.driver.OracleDriver>,
 > expected: null

Yes it was failing for me as well called from Maven with pom.xml as is.
So i moved the oracle dependency from oracle profile to default profile.

I hadn't time to investigate why it is failing and not just skipping as it was implemented
and expected in the change you've referenced.

Patch-set: 4