Fix conditional skipping of tests for unavailable databases
Change If4ae2aa added assumeNoException calls in the test classes'
@Before annotated methods to skip testing for databases that are either
not installed or not properly configured on the system.

However, the assumeNoException method only causes tests to be skipped
when it is called from a method annotated with @BeforeClass.  When it
is called from within a method annotated with @Before, the tests still
fail.

Refactor the checks into methods annotated with @BeforeClass.

Change-Id: I779f48e062779d73f8d7a7aec32a6e933bfa6e13