Fix conditional unit test skipping
Change If4ae2aaec added assumeNoException calls in the test classes'
@Before annotated methods to skip testing for databases that are either
not installed or not properly configured on the system.

I29ab37e1c broke that, by moving database cleanup operation from setUp()
to tearDown(), but not taking into consideration, that when test was
skipped, executor wasn't properly instantiated and it come to NPE in
tearDown() method. Failures in tearDown() then caused test to be marked
as failed.

Change-Id: I6964feedbe44ad9849c4feaa2565d0089a84509a