Restructure the byte code generation code for the LIMIT support
It was hard to follow the chain of if statements which differentiate
between the following cases:

1 SQL dialect supports LIMIT and LIMIT ? was used
2 SQL dialect supports LIMIT and LIMIT N was used
3 SQL dialect doesn't support LIMIT and LIMIT ? was used
4 SQL dialect doesn't support LIMIT and LIMIT N was used

For the cases 3 and 4 we generate code which does:

  ps.setMaxRows(...)

For the case 1 we generate code which does:

  ps.setInt(...)

For case 2 there is nothing to do as LIMIT N is part of the SQL
query text and the dialect supports limit.

The code generator used to generate both commands:

  ps.setInt(...)
  ps.setMaxRows(...)

for the case 1. This wasn't necessary and is also fixed with
this change.

Change-Id: I788473ec44b881087b5c0e264d4fab2d8bce0db4