Add SqlDialect method to drop an index
Only MySQL has a custom implementation of this.

At this point it is easiest to only add a drop method, not an add
method. There are multiple ways add could be implemented, e.g.
automatically creating indexes when scanning over @Query fields.
However, this change is primarily to help Gerrit Code Review migrate
away from gwtorm, for which dropping indexes is more germane.

Change-Id: I2eacad107d8f6e78c77f43f30ca22a62d3a65f02