Make Schema and StatementExecutor interfaces AutoCloseable
We target Java 7, so this should make things nicer. We do not make
ResultSets AutoCloseable because they are implicitly closed when
iteration is finished, and close should otherwise be called manually;
they aren't really used from ARM blocks.

Override the existing close methods so we can remove the exception, as
the existing close implementations don't throw.

Change-Id: Ibc628eb1d7694e52078bacef4866a2ec55b369d2