Update patch set 4
Patch Set 4:

for info:

in postgresql:
"Adding a column with a non-null default or changing the type of an existing column will require the entire table and indexes to be rewritten. As an exception, if the USING clause does not change the column contents and the old type is either binary coercible to the new type or an unconstrained domain over the new type, a table rewrite is not needed, but any indexes on the affected columns must still be rebuilt. Adding or removing a system oid column also requires rewriting the entire table. Table and/or index rebuilds may take a significant amount of time for a large table; and will temporarily require as much as double the disk space."

iow the reorg is implicit and immediate.

Oracle:
"CHECKPOINT Specify CHECKPOINT if you want Oracle Database to apply a checkpoint for the DROP COLUMN operation after processing integer rows; integer is optional and must be greater than zero. If integer is greater than the number of rows in the table, then the database applies a checkpoint after all the rows have been processed. If you do not specify integer, then the database sets the default of 512. Checkpointing cuts down the amount of undo logs accumulated during the DROP COLUMN operation to avoid running out of undo space. However, if this statement is interrupted after a checkpoint has been applied, then the table remains in an unusable state. While the table is unusable, the only operations allowed on it are DROP TABLE, TRUNCATE TABLE, and ALTER TABLE DROP ... COLUMNS CONTINUE (described in sections that follow)."

so there is not a 'reorg per say, but still, the table become unusable until every row is touched to reflect the change.. same difference.
do that casually on a large table, and you're in for a world of pain.

with DB2-luw alter add column (even with default) is not a reorg-pending operation... most alter 'change' column and alter drop column put the table in reorig pending state....
The idea is that you may do more than one of these.. and doing a reorg 'automatically' on the fly is not compatible with normal operation on a serious production database (which db2 is designed for)

Also bear in mind that Drop Column is incompatible with Change Data Capture (option to trigger extra logging needed for replication).


DB2 z/OS allow these columns change without reorg-pending state.
the reason is that on (fairly) recent version, the on-disk format of a row contain a byte that identify a 'row' version...
Internal data structure contain the description of all old versions, allowing db2 to convert old row to the most current format on the fly as needed.

Alter add column is a special case in the sens that detecting the absence of row at the end of a physical record is trivial, so any reader could complement these on the fly, knowing the most recent description (which is in the catalog, including the eventual user-specificed default value)

Bottom line is: supporting db2-luw for 'use' is fine.. but supporting properly gerrit upgrade with schema changes is going to be much more tricky.

Patch-set: 4
Reviewer: Gerrit User 1015382 <1015382@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0