Add support for Apache Derby database
Renaming of a column isn't supported as long as a constraint exists on
the column. Currently a CHECK constraint is created by gwtorm for all
fields with type SqlBooleanTypeInfo. When for a schema migration a field
of type SqlBooleanTypeInfo is renamed the schema migration will fail
with the current Derby dialect implementation. See also
https://db.apache.org/derby/docs/10.11/ref/rrefsqljrenamecolumnstatement.html
for more information about the restriction of the RENAME COLUMN command.

Change-Id: I2dd50e3122f5ca59c6712dec49cdd004c84173a4
Bug: Issue 3441