Update patch set 2
Patch Set 2:

> > Tested only on H2 and Apache Derby.
 > 
 > Is there likely to be any impact on other DBs?  I.e. should we make
 > some effort to explicitly test those too?

I don't expect any surpises here. The same create sequence statement code path, that only was used for account sequence before this change, would be activated to every sequence, to explicitly set "start with" to 1. It wouldn't hurt to create new test site for PostgreSQL and/or MySQL and check that it still works, though.

Patch-set: 2