Add dialect for SAP HANA
A new database dialect is added to support SAP HANA [1].

The dialect uses the column store of the HANA database and hence
needs to create tables using a 'CREATE COLUMN TABLE' statement. A new
method SQLDialect#getTableTypeSql is introduced to support this: its
return value defaults to "TABLE" but is overwritten by "COLUMN TABLE"
in the HANA dialect.

[1] http://help.sap.com/hana_appliance/

Change-Id: Ib733610cb826d17d72defd71a9e2c5f0fc7ce154
Signed-off-by: Adrian Goerler <adrian.goerler@sap.com>